Consider page size when sizing tlb_locality
DEF_TLB_LOCALITY is meant to be 64 pages, but is defined as
a size (64*4096). ppc64le has a 64kB page size, so this is only
4 pages.

Change DEF_TLB_LOCALITY to be in pages, and use getpagesize() to
convert it to a size. Without this patch, the numbers on ppc64le
are impossibly good because of hardware prefetching.