Remove alloc_arena_shm.
This is a first step towards revamping hugepage support. Once this is
done, this function will no longer be used - so we can clean it up.

A couple of problems we're trying to solve:

- Even if the '-H' flag wasn't specified, the test may in fact use
  hugepages anyway, via Linux's THP feature.
- The '-H' flag is just a toggle, it only lets us use a single hugepage
  size (Hugepagesize from /proc/meminfo), whereas systems commonly
  support different hugepage sizes (e.g., x86_64 might support all of
  4k, 2m, and 1g pages).

So, going forward, the plan is to add a new argument, '-p', which lets
the user pick a page size explicitly. We'll modify the arena allocation
function to ensure we really do use this page size.

For the native page size, we'll madvise() to tell the kernel not to back
the area with THPs. For hugepages, no need to use shmget(), we can just
pass MAP_HUGETLB into mmap(), along with an appropriate page size flag
(e.g., MAP_HUGE_2MB).