Plumb page size through arena allocation, disable THP.
Another step towards revamping hugepage support.

Modify the arena allocation function so it takes the page size as an
argument. For now, this is only ever the native page size on the machine
(e.g., 4k).

Setup a helper which returns mmap() flags based on the page size. For
now it only supports the native (non-huge) page size, and so it just
returns zero - in future commits, it will return appropriate MAP_HUGETLB
flags.

Finally, after we mmap() the arena, if we're using the native page size,
explicitly madvise() the area, to instruct the kernel not to use THPs to
back it. This lets us avoid accidentally using hugepages when we didn't
mean to.