Check transparent hugepage state, and change it if needed.
The previous commit is slightly incomplete. MADV_HUGEPAGE will return
"success" but silently do nothing if e.g. hugepages are disabled, or set
to "defer" on the system.

It seems useful to avoid a case where we run a benchmark thinking that
THPs are being used, while in reality they are not. To solve this, read
the current state from sysfs before allocating the arena. If the state
is such that hugepages wouldn't be used, try to change the state as a
convenience.