Change Multiload load_stream_triad to 64B align
In original implementation, multiload's load_stream_triad and
load_stream_triad_nontemporal_injection_delay is aligned by 16 byte
addresses. As this is not aligned with the cache line size, it results
in additional work when accessing memory which impacted the results of
the benchmark.

To fix this change, we are proposing to change the alignment from 16
bytes to 64 bytes.